L189: Improving the paging experience
Replicating componentDidUpdate with hooks proved to be a bit of a
challenge. In the end a combination of useEffect and useRef worked. I
just need to remember to pass an updater function to the setter returned
by useState instead of the new state, otherwise there will be an
infinite loop.